+++ packages/storage/src/abstract-disk.service.ts
